requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
einops>=0.6.1
timm>=0.9.2
lpips>=0.1.4
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scipy>=1.10.0
psutil>=5.9.0
jupyter>=1.0.0
ipywidgets>=8.0.0
